WIP  - Move `Platform` `*Shell` and `Sway{Sock,Error}` to common/nwg_classes.cc  - Save reference to window in `Platform` instead of taking it as an argument
