+++ grid/grid.cc
@@ -12,11 +12,6 @@
 #include <sys/stat.h>
 
 #include <charconv>
-#include <variant>
-
-#ifdef HAVE_GTK_LAYER_SHELL
-#include <gtk-layer-shell/gtk-layer-shell.h>
-#endif
 
 #include "nwg_tools.h"
 #include "nwg_classes.h"
@@ -47,93 +42,6 @@ Options:\n\
 -c <name>        css file name (default: style.css)\n\
 -l <ln>          force use of <ln> language\n\
 -wm <wmname>     window manager name (if can not be detected)\n";
-
-struct GenericShell {
-    Geometry geometry(Gtk::Window& window) {
-        Geometry geo;
-        auto display = window.get_display();
-        if (auto monitor = display->get_monitor_at_window(window.get_window())) {
-            Gdk::Rectangle rect;
-            monitor->get_geometry(rect);
-            geo.x = rect.get_x();
-            geo.y = rect.get_y();
-            geo.width = rect.get_width();
-            geo.height = rect.get_height();
-        }
-        return geo;
-    }
-    void show(Gtk::Window& window) {
-        window.show();
-        window.set_type_hint(Gdk::WINDOW_TYPE_HINT_SPLASHSCREEN);
-        window.fullscreen();
-    }
-};
-
-struct SwayShell: GenericShell {
-    SwayShell(Gtk::Window& window) {
-        window.set_type_hint(Gdk::WINDOW_TYPE_HINT_SPLASHSCREEN);
-        window.set_decorated(false);
-    }
-    SwaySock sock_;
-    void show(Gtk::Window& window) {
-        // We can not go fullscreen() here:
-        // On sway the window would become opaque - we don't want it
-        // On i3 all windows below will be hidden - we don't want it as well
-        sock_.run("for_window [title=~nwggrid*] floating enable");
-        sock_.run("for_window [title=~nwggrid*] border none");
-        window.show();
-        // works just fine on Sway/i3 as far as I could test
-        // thus, no need to use ipc (I hope)
-        auto [x, y, w, h] = geometry(window);
-        window.resize(w, h);
-        window.move(x, y);
-    }
-};
-
-struct LayerShell: GenericShell {
-#ifdef HAVE_GTK_LAYER_SHELL
-    LayerShell(Gtk::Window& window) {
-        // this has to be called before the window is realized
-        gtk_layer_init_for_window(window.gobj());
-    }
-    void show(Gtk::Window& window) {
-        window.show();
-        auto gtk_win = window.gobj();
-        std::array edges {
-            GTK_LAYER_SHELL_EDGE_LEFT,
-            GTK_LAYER_SHELL_EDGE_RIGHT,
-            GTK_LAYER_SHELL_EDGE_TOP,
-            GTK_LAYER_SHELL_EDGE_BOTTOM
-        };
-        for (auto edge: edges) {
-            gtk_layer_set_anchor(gtk_win, edge, true);
-            gtk_layer_set_margin(gtk_win, edge, 0);
-        }
-        gtk_layer_set_layer(gtk_win, GTK_LAYER_SHELL_LAYER_TOP);
-        gtk_layer_set_keyboard_interactivity(gtk_win, true);
-        gtk_layer_set_namespace(gtk_win, "nwggrid");
-        gtk_layer_set_exclusive_zone(gtk_win, -1);        
-    }
-#endif
-};
-
-struct Platform {
-    Platform(Gtk::Window& window, std::string_view wm): shell{GenericShell{}} {
-#ifdef HAVE_GTK_LAYER_SHELL
-        if (gtk_layer_is_supported()) {
-            shell.emplace<LayerShell>(window);
-            return;
-        }
-#endif
-        if (wm == "sway" || wm == "i3") {
-            shell.emplace<SwayShell>(window);
-        }
-    }
-    std::variant<LayerShell, SwayShell, GenericShell> shell;
-    void show(Gtk::Window& window) {
-        std::visit([&](auto& shell){ shell.show(window); }, shell);
-    }
-};
 
 int main(int argc, char *argv[]) {
     std::filesystem::path custom_css_file{ "style.css" };
@@ -363,7 +271,7 @@ int main(int argc, char *argv[]) {
     MainWindow window(execs, stats);
     window.set_background_color(background_color);
     Platform platform{ window, wm };
-    platform.show(window);
+    platform.show();
 
     gettimeofday(&tp, NULL);
     long int images_ms  = tp.tv_sec * 1000 + tp.tv_usec / 1000;
